requirements.txt
torch
torchvision
dgl-cu102
pyyaml~=5.3.1
numpy~=1.18.5
yaml~=0.2.5
scipy~=1.5.2